chức năng đặt hàng

+++ src/main/java/com/cdtn/computerstore/dto/cart/response/CartItemDetail.java
@@ -13,7 +13,7 @@ public class CartItemDetail {
     private Long productId;
     private String productName;
     private String productImageMain;
-    private Long itemQuantity;
+    private Integer itemQuantity;
     private Long productLatestPrice;
     private Long totalPricePerProduct;
 }
